rag-pipeline/requirements.txt
# RAG Pipeline - Free Tier Dependencies
# All packages have free tiers available

# Core
streamlit>=1.28.0
pyyaml>=6.0

# Qdrant Vector Database
qdrant-client>=1.8.0

# Embeddings (Free - sentence-transformers)
sentence-transformers>=2.2.0

# LLM - Google Gemini (Free Tier Available)
google-generativeai>=0.4.0

# LangChain for agent framework
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.0

# Optional: FastAPI for REST API
fastapi>=0.104.0
uvicorn>=0.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
